Transition from passing all generic parameters as pointers to passing by-value where possible.
This solves some issues where due to the semantics of compound literals,
temporaries can consume extra stack space in the caller and linger until
the end of the function. The new implementation allows the compiler to
optimize in more cases.

Simplify the calling convention for custom type ops and eliminate the
ST_Indirect flag.

Variants are now fully fledged stypes built on top of generic arguments.