Change string typedef back to a nonexistent struct pointer
This removes the convenience of simply passing a C string in place of a
cx string variable, but provides much greater static type checking and
protection against accidentally passing a cx string to a function that
is expecting a C string.

To use C strings they can be explicitly cast to (string).