Add a second VFS lock for "major" operations that can invalidate cache data
This allows _vfsFindMount to safely drop the cache lock when calling
the provider's stat(), preventing a deadlock in the case of VFSVFS being
used to loop back to the same VFS (and is the reason such loopback mounts
were prohibited and triggered an assert). VFSVFS loopback mounts are now
allowed.

It turns out that _vfsFindCIHelper was already *unsafely* dropping the
lock there for a split second, which might have theoretically caused a
crash in extremely rare circumstances, and then only when a case-sensitive
filesystem was mounted on a case-insensitive VFS. That potential issue
is now fixed.