Add strref pseudo-type for borrowed string references
This pseudo-type has the same stype ID as string, but is typedefed as
a pointer to a const. That allows abuse of compiler const semantics to
enforce the rule that strings passed as functions parameters (now strrefs)
cannot be assigned to string variable or passed as an output string
to a function that takes a string*.

Also add a string(name) macro to be used for declaration of string
variables that properly initializes it to 0.