The Big Giant Iterator Revamp
Standardize on an init->next->finish cycle, where init fills out an
iterator structure that is usually stack allocated and sets it to point
to the first element, next steps through the elements, and finish frees
any resources involved in the iteration.

Many iterators also provide a 'valid' function or macro, which makes
things more convenient for 'while' and 'for' loops.

Add foreach construct that takes advantage of the newly consistent
semantics.