Make object static-cast macros paranoid about possible braindead compilers
Taking the address-of the identity field should be safe, especially on
the left side of a comma where it will be optimized to a no-op anyway.
Despite that, be extra paranoid in case some broken compiler decides
it needs to dereference the (possibly NULL) pointer to get the address.
This check will also be optimized away by any sane compiler so it costs
nothing.