The great sarray revamp
sarray was designed for convenience by having the array handle look like
a plain old C array. While handy, it's very hard to do any kind of
sensible type checking and could lead to hard to detect mistakes involing
wrong pointer indirection level.

Redesign sarray to include a type safety mechanism. A small bit of
convnience is sacrified as the array contents must be accessed through
the .a member of the union that replaces the array pointer. sarray
types must also be predeclared by using the saDeclare, daDeclarePtr,
or saDeclareType macro.

saCreate was redesigned as saInit to follow the same paradigm as iterators
and other similar objects. Assignment-based initialization is no longer
possible due to the compiler not knowing the individual types are
identical.