Rework container interface to use references rather than handles where possible
Functions that only read from the containers and cannot alter the
container itself no longer require a pointer-to-a-pointer, but just a
direct reference using the container's base type.

Functions that can modify the container continue to require a handle
(pointer to the container) as the container's address may change.

One notable semi-exception to the rule is that saFind and htFind still
require a &handle, because they may alter the container if the Remove
or Destroy function flag is present.

This makes the interface consistent with the string API.