Container reference counting
sarrays and hashtables are now reference counted objects that use the
new refcount module with single-threaded optimizations to avoid atomic
operations in simple cases.

saDestroy and htDestroy are now saRelease and htRelease, which are
drop-in replacements for all existing uses.

Added saAcquire and htAcquire.

Stype behavior changed for these two containers; adding one to an array
or hashtable will now acquire a reference rather than making a copy. Be
aware that this changes the behavior for existing use of nested
containers.

Added saClone and htClone for making a full copy of containers.