Remove function flags
It was a neat preprocessor trick, but ultimately not worth the effort to
sometimes save a couple of characters typing at function call time.

The biggest downside was that it made the API harder to master and
required the function variants to be committed to memory to use them
effectively. The standard way of using flags in function calls is well
understood and things like Intellisense hints can make it easier to
discover the different flags, which did not work with the macro system.