Logging system revamp
logBatchBegin() / logBatchEnd() can now be nested to make writing
subroutines that use them easier. Only when the outermost logBatchEnd()
in a thread is called is the batch submitted to the log writer.

Log batches are stored as linked lists for constant-time insertion.

Log batches are inserted as a single pointer into the log buffer,
significantly reducing multithreaded contention when large batches are
in use.

Log ringbuffer growth is capped to a maximum size.

Rework buffer interaction between producer threads and log writer thread
to make it more efficient and reduce unnecessary waiting. Yield to other
threads as needed to reduce CPU spinning and increase the chances of a
successful compare-and-swap.

Add logic to log buffer insertion to give up after a set number of retries,
to ensure that it can't spin forever when under extreme contention.

If the ringbuffer is full or thread contention is too high, a per-thread
overflow list is used to temporary store log entries and retry them the
next time something is logged. Only if both the ringbuffer and the
overflow buffer are full are log entries discarded. Losing log entries
can only happen under extreme circumstances, and a special log entry
is generated to note this exceptional event.