Add rwlockDowngradeWrite
This function allows a caller who holds the write lock to atomically
downgrade it to a read lock. The main reasons to do this are:

* To avoid a potential wait if the read lock needs to be immediately
reacquired, but another writer jumped in and grabbed it in the brief
interval when it was unlocked.

* To guarantee that in a situation where the write lock is dropped
and the read lock is reacquired, no other thread can make changes
to the data protected by the lock in the interim.