Replace alertable futexes with UI Events
Sigh, the alertable flag on NT keyed events didn't actually do anything
useful, apparently it only affects APCs and certain I/O operations.
Hooray for undocumented APIs.

Rip out all of the alertable futex stuff since it doesn't work, and
replace with a mechanism for Events to use a platform-specific
kernel object instead of a futex, if the Event is created with a flag
to allow it to be woken by UI input. Only makes sense on Windows
currently.