Remove switchable allocator support
Previously, there were reasons to sometimes use one allocator over
another.

With the latest update, mimalloc has improved enough that there
are no longer compelling reasons to maintain this support. It now works
on 32-bit windows as well as WebAssembly, and while jemalloc is still
slightly faster on FreeBSD and arm64, the performance gap has closed
significantly and there is not enough of a benefit to justify the added
complexity and maintenance overhead.

This also opens the door to integrating the allocator more tightly with
cxmem and adding advanced features built around mimalloc's multiple
heap support.