Fill out fmtopts array only when needed.
fmtopts is not used by any of the built-in formatters, it's only
for the extensible object formatting by the Formattable
interface.

Don't fill out the array when it's not needed, instead move that
responsibility to the object format string parser as that's the only
context where it will actually be used.