strFormat performance optimizations
Revamps to strFormat in order to minimize heap allocation and
make more use of the stack. Temporary strings used during parsing
and formatting are completely replaced with shared stack-allocated
strings that are reused.

Preallocates an estimate of buffer space for the result to reduce the
amount of reallocation and expansion as the string is constructed.

Caches length of format string to avoid repeated cstrLen calls.

Altogether these optimizations increase the average-case performance
of strFormat by about 21%.