Streamline foreach
Rewrite the foreach macro using a trick involving nested
single-statement for "loops" to declare variables with limited
scope. This eliminates the awkward endforeach macro that
used to be necessary.

Despite the scary looking nested loops, this actually generates
extremely clean assembly on any reasonable optimizer setting.

Note: This is a breaking API change.