Standards compliance pass on pblock
Fixed a couple of minor breaks with the C standard. Even though in
practice it works with the ternary operator in just about every
compiler, strict C spec adherence demands it only be called in 1 of 4
very specific ways. Rewrote the case labels to make this possible, and
improved documentation.

Also marked a couple of local variables in the unit test volatile that
are accessed by a pblockReturn expression, which occurs after a longjmp.