Scrap pblockReturn
It's sad, but there just isn't a way to make this work across all
supported platforms.

The longjmp back to the return site is borderline standards-violating.
It can be argued that it technically doesn't jump back to a frame that
has been exited because all of the longjmps take place within a single
function and thus in the same stack frame. However, it does jump
"forwards" by going back to a deeper level of nesting after already
having longjmped out earlier.

While the standards compliance is debatable, emscripten/WebAssembly's
longjmp implementation can't handle this and it causes the eventual
return to fail to return to the caller and execution instead goes to
unexpected places.

Other methods attempted involve saving the return value in order to do
the return itself at the top-level pblock, but this is not viable
because the return value type can vary from function to function or even
be void.

So, while it's unfortunate to lose the feature, it's not a huge burden
to work around by having a local ret variable or similar. Only using
longjmp to go "backwards" keeps the code standards-clean and
working across all supported platforms.