Rework ptry to remove _pblock_unwind_prev
Since it's not permitted to mix and match pblock and exceptions*, the
logic can be simplified by removing some of the pblock/ptry integration.

In particular, the TLS handler stack is now always used, ignoring
the current _pblock_unwind_top at throw-time. This has the advantage
that it no longer needs to be hidden inside of catch blocks for throwing
to work correctly there. That in turn eliminates the need for
_pblock_unwind_prev entirely, which means one less local variable is
needed for every pblock.

* It's not clear if that's supposed to be supported in P99, but it
doesn't actually work to mix and match the equivalents there either.