Add Semi-Structured Data trees
Semi-Structured Data trees (ssdtrees) make use of containers like
hashtables and arrays to represent a hierarchical structure of values,
usually mapped as name-value pairs.

This is a structure it was possible to build manually, but ssdtree
standardizes it in a thread-safe way.

This is also the foundation for a json-compatible structure.