Remove threadsafe support from streambuf
This seemed like a good idea at first, but during implementation of
ThreadedStreamBuffer it became clear that the locking design was
woefully insufficient to be able to tie StreamBuffers together across
threads without one thread always being blocked.

Rather than increase complexity to make that work, refocus the design
on the core goal, which is coroutine-style streaming within a single
thread.

This significantly reduces code complexity and removes the need for a
couple of fairly ugly hacks.

Slight API change in that the buffer is now reference counted in order
to handle finishing the producer/consumer from inside the callback. A
user of sbufCreate must call sbufRelease on the returned pointer before
discarding it. This change makes the design more robust as it's harder
to accidentally end up with a dangling pointer.