Comment support for cxobjgen
This one has been a long time coming.

Scan comments (// style only!) in sidl files and reproduce them in the
generated header file so that IDEs like Intellisense can show them in
tooltips.

Works for class data members, interface methods, and the #defines
generated for calling class methods.

Automatically adds the function prototype as a comment for function
call wrapper macros so that users of the macro don't have to remember
or guess the parameter types.