Standardize on uint8* for pointers to byte buffers
char being signed makes it an inherently inconvenient type for doing
any kind of byte manipulation, so use unsigned chars for generic byte
buffers as well as string buffers.

This works well for the string type. It is inherently UTF-8, which
may often contain bytes in the 0x80 - 0xFF range. The only exception
is C-style strings which remain char* or const char* for
compatibility.