New hashtable implementation
This replaces the 'textbook' hashtable implementation with a compact
sparse hashtable that uses a series of chunked storage arrays that are
densely packed and a sparse lookup table that must only store the small
index into the storage arrays.

This results in both lower memory usage and better cache locality.

As an interesting side effect, the new implementation also preserves
insertion order and thus maintains a stable iteration order over the
lifetime of the table.