Add type conversion to stype
This is an stype mechanism for converting from one type to another. It
is the responsibility of the source type to know how to convert to a
given target type (or something close enough it can be handed off).

As a runtime system this can also be used with stvars whose contents
may not be known at compile time.

Fully integrated with STypeOps and includes an object interface for
conversions involving objects.