Clarify expectations for use of stConvert
Resolve a conflict between st* functions wanting to be a low-level
interface capable of dealing with uninitialized memory, and the object
conversion wanting to be a high level interface capable of efficiently
reusing memory or feeding into a stack-allocated string buffer, e.g. for
strFormat.

Add some comments clarifying the expectations of how each should
be used.