Add SSD array import/export functions
Useful for moving data in bulk in or out of array nodes.
Import/ExportArray operate on arrays of stVars.

Import/ExportTypedArray operate on arrays of a single type.
Strict mode for export requires the the array node only contain
values of the given type, non-strict mode will skip over values
with an uninteresting type.