SSD Iterator locking tweaks
No longer holds the lock for an indeterminate period of time if
transient locks are being used. Now locks and unlocks around each
use of the iterator in that case.

Explicitly locked iterators lock when needed but do not unlock -- it is
up to the caller to manage the lock.