Add ST_Equality flag to stCmp
This flag can be used to indicate that the caller is only interested in
equality and is not trying to sort the operands. It can be used in
contexts like testing if two stvars that may contain objects are equal,
which normally would not be possible if the objects don't implement
Sortable or without special casing.

Now stCmp can be used with ST_Equality and it will fall back to a
pointer comparison in that case.