Streambuffer streamlining
Streambuffer now automatically finishes the side that operates on
callbacks (producer side in pull mode, consumer side in push mode).

This means that the callback doesn't need to worry about calling
sbuf*Finish; it will happen automatically when the driving side
finishes.

It may still be called early -- once finished that side will not receive
any more callbacks so it is safe.