Further refine hashtable implementation
Key and value storage chunks are now managed as separate arrays. This
has a few effects:

  * Previously there was a minimum element size of 16 bytes (128 bits),
    wasting space if the key+value length was less than 16 bytes.
  * Keys and values are now tightly packed and have no minimum size.
    This reduces memory usage when using small keys and values.
  * Avoids alignment issues on ARM platforms when using odd
    combinations such as int8 keys with pointer values.
  * Better cache locality improves hashtable iteration speed.
  * Hashtable element deletion speed is improved by approximately
    5%.
  * No measurable performance hit to insertion or lookup.

It is now legal to use a value type of "none" in order to create a
hashed set rather than a key/value mapping. In set mode, use the keys
as the set members while leaving value as none / NULL.

In set mode, the hashtable will not waste space or time storing
nonexistent values.