Rework asserts again
Partially revert previous changes. *Assert returning a value doesn't
really make sense, since the point of asserts is to not evaluate the
expression in release builds.

Assets now are all explicitly cast to void to prevent them from being
used in expression context.

Add dbgVerify and devVerify with slightly different semantics. This
family always evaluates the expression and coerces it to a boolean. They
can be used in an error-handling if() statement which will be checked
even in release builds, but only asserts on failure in debug or
development builds.