xalloc revamp phase 1
Start reworking xalloc. Now that mimalloc is the only supported
allocator, this opens the door for a much more in-depth set of features.

Phase 1 splits the core functions (malloc, realloc, and free) to exist
in cxmem for third-party library support while moving the rest to the
main cx library for future expansion.

BREAKING CHANGE: xaResize now takes a pointer-to-pointer and replaces
it with the reallocated pointer on success, and returns a boolean.

BREAKING CHANGE: xaExpand removed, as it was a no-op with mimalloc.
Code that wants the ability to expand in place should use the new
xaOptSize function to determine the optimal size to allocate. Otherwise,
in secure mode, expansion in place was always impossible because
due to secure padding, mimalloc's usable size function returns the exact
size the user allocated even if it was placed in a larger size class.