Use object libraries to compose cx library
Instead of producing a bunch of static libraries that have to be
linked in, use object libraries to produce a single static library
(cx.lib / libcx.a). This eliminates the circular dependency problem
with third-party libraries and the allocator, which will be necessary
for the next phase of the xalloc revamp and folding cxmem into cx.