Handle out of memory conditions in logStr
Log messages are silently dropped if there is not enough memory to
allocate a log record, and may be shunted into the overflow list if
there isn't sufficient memory and a buffer expansion is needed.

This is needed so that the allocator can use the log module to record
warning messages, while still being able to recover from an OOM
situation.