Add out-of-memory handling to xalloc
This system allows user-provided callbacks to be invoked when the
allocator does not have enough free memory to satisfy a request.

These callbacks can free unimportant memory such as cache data and
allow the allocator to try again.

If there is still not enough memory even after all the callbacks are
finished, the allocator will either assert and terminate the program if
it's a required allocation, or return NULL if it's an optional one.

There are three phases of 'effort' with defined rules for what the
callbacks should try to do in each phase. Low-tier optional allocations
may not invoke all (or any) of the phases, while required allocations
always exhaust all phases before giving up and terminating the program.