Use a static default log category rather than dynamically allocated
This prevents a condition where the category in a log message could be
NULL rather than LogDefault for the first message if logStr/logFmt was
called before any other logging functions.

In practice this doesn't really matter because it can only happen if
the logging system hasn't been initialized yet, which means there are no
destinations registered, so the message with the incorrect category goes
nowhere.

Despite that, go ahead and guarantee that the category will never be
NULL in order to be strict about the invariants.