Make rwlock basic acquire never fail
These can technically fail even with no timeout if they hit the limit
on number of readers or writers that can be waiting on the lock at once.
These are large limits and shouldn't be hit in any kind of sane design.

As such, rework the basic read/write acquire to return void and be
guaranteed not to fail, triggering a fatal error if one of those limits
are exceeded. This allows most locking code to not have to worry about
handling failure cases, which would make proper usage much more
complex.

Callers who need the old behavior can instead call
rwlockTryAcquireTimeout directly, even with a timeForever timeout, so
long as they are prepared to handle the failure case.