cxobjgen support for guaranteed object creation
Factories and init methods can be annotated with [canfail] to indicate
that they may fail.

Init methods that do not have [canfail] are required to return true.

cxobjgen automatically calculates if a factory can fail or not (and
whether it may return NULL) depending on whether the class's init
function or any parent/mixin class initializers can fail. Factories that
have an independent possibility of failure may also be directly
annotated with [canfail].

This helps the static analyzer better understand whether or not code
that creates objects needs to check the return value upon object
creation or not.