Add deferred logging
This is a temporary buffer that can be registered as a log destination
early to collect program logs that are generated before a logfile or
other heavy destination is set up.

The logRegisterDestWithDefer function registers the real destination
while atomically removing the placeholder defer buffer and sending
all the logs it collected to the real destination.