Improve FS/VFS Search API
fsSearchInit / vfsSearchInit now return true if the path exists but is
empty. Previously it was not possible to distinguish between a path
that could not be accessed and one that was simply empty with no
files.

Callers should be sure to make use of the *SearchValid() functions
before trying to access the iterator, as an initialized search may not
necessarily contain any results! Most users, including the foreach
wrapper, already did this.

VFS providers are also now required to implement a searchValid
method.