Remove nonstandard config types and compiler flags
For better integration into other projects. The project that includes
cx can define CX_BUILD_TYPE as one of Debug, Dev, or Release
(or a generator expression that results in one of those strings)
to override the behavior for the standard CMake configurations.

Inherit all compiler flags and optimization settings from the parent
project. Warnings-as-errors are enabled specifically for the cx project,
but not for the third-party sources.