Add Task Queue
Task Queue is a system for object-oriented programs to run discrete
units of work (tasks) on a pool of worker threads. The queue is backed
by the new lock-free pointer-ring buffer and supports full MPMC
usage, though it is also quite capable and performant in MPSC and
SPMC scenarios.

It supports automatic pool size scaling, a mechanism to defer tasks
either until another task successfully completes or to be run after a
specified delay, hassle-free memory and object lifecycle management
with cleanup deferred to an independent thread to not tie up workers.

It also includes a minimal Basic Task class that can be used for lighter
weight applications that don't need the full functionality, and a simple
callback-based interface for non-OO usage.