Improve queue responsiveness
Add ability for tasks to specify an event to signal AFTER their
success/failure is set. This is useful for preventing races.

Automatically wake up an extra thread when a worker completes a run
and processed some tasks. If the queue is empty it will go right back to
sleep but if it's not, it helps them scale up faster.