Add convenience functions for pointers and objects in stvar lists
This is more syntactically convenient for some common use cases in
variadic functions.

Note that it cannot distinguish between a NULL pointer in the list and
a failure because the end of the list was hit without finding a value
of the correct type, and should only be used in cases when that is
acceptable.