Add closure and closure chain types
These are first class stypes that represent a function instance that
has some associated variables (cvars) stored alongside it. The
cvars take the form of a list of variants, much like a variadic
function.

They can be used for callbacks that store some arbitrary extra
data which is automatically cleaned up when the instance
is destroyed.

Closure chains are a thread-safe structure that links multiple
closures together. An example of how this could be used is an
event mechanism where one or more interested parties register
a callback in a closure chain.