Task defer improvements
Add a TASK_Deferred state.

Allow tasks to be deferred indefinitely by setting their defertime to
timeForever.

Add advance() method to remove a task from the deferred list and
make it immediately runnable.

Add convenience macros to check for common task states.

Destroy oncomplete callbacks as soon as a task succeeds or fails. If
a completed task is re-used it will not have the callbacks registered,
making them one-shot callbacks in that scenario.

Show details of time-based deferred tasks in monitor output.