Fix defer list comparison function
Wasn't an issue before taskAdvance was added, when it was only
used for sorting the list. But with the advance queue implementation,
find and remove needs to be able to identify the specific task.
Otherwise it might remove the wrong one causing deferred tasks
to be lost and potentially a task to be run more than once on different
workers.

Use the pointer value as a secondary key after the next run time.