MTask performance improvements
Reduce overhead by using the task advance mechanism instead of
running the full cycle every time a task completes. This batches them
up if multiple tasks complete concurrently and is in testing proved to
be considerably more efficient than even using mutexTryAcquire
because the pointer-ring queue is so fast.

Keep a counter of how many completed tasks need to be processed to
ensure that the task continues to be run until all are accounted for.