Disable autogeneration of objstdif
This resolves a circular loop of cxobjgen depending on itself when the
ninja generator is used.

The real fix is to add DEPENDS_EXPLICIT_ONLY to the custom commands,
but that requires CMake 3.27 which is quite new and not in several
popular distributions yet.

See https://gitlab.kitware.com/cmake/cmake/-/issues/17097