Start implementing nullability qualifiers
To complement the use of SAL on Windows/MSVC, start adding nullability
qualifiers to pointers. This commit covers the string module.

This allows clang in particular to do better static analysis.

Give in and change clang-format to align pointer asterisks on the left,
since that makes more sense for the placement of the nullable / nonnull
qualifier.