Add natvis decoder for cx string type
This allows Visual Studio (and VS Code using either MSVC toolchain,
or gdb with a customized launch.json file) to decode the packed string
type in the debugger and show the actual string contents.

Fully supports all string features. Ropes cannot show the fully
assembled rope but do show the rope references which can be followed
to see the pieces.