Add GDB version of natvis file
It turns out that the GDB MI support for natvis is very lacking and
missing many features. It also has some important differences in the way
typedefs are handled (or not).

To that end, create a separate natvis file for use with GDB. Due to the
lack of intrinsics, everything has to be unrolled to some really nasty
expressions, but this can be done in a mostly automated form.

This file can be used by putting the following in launch.json:

    "visualizerFile": "${workspaceFolder}/cx/cx-gdb.natvis",
    "showDisplayString": true