cxobjgen: Don't fight with clang-format
Implement for function definitions in implementation files so that if
the prototype cxobjgen wants to write differs from the original text
in the file only in whitespace, keep the original text. This prevents it
from reverting clang-format's splitting of parameters onto separate
lines.

Add comments to disable the formatter in purely autogenerated
sections so it doesn't do things like reorder the #includes.