Revamp TaskQueue
Switch to modular design to make the queue more flexible to meet
different needs.

Ability for manager to run in-worker rather than on a dedicated thread.
This spreads managers duties among the workers once the are done with
a run of the queue, or once every 16 tasks at most.

Add manual mode for a task queue that does not use a separate thread
pool but is manually ticked; could be used as part of an event loop.

Add NoComplex queue type with simplified worker / manager logic, but
that cannot run only simple tasks that succeed or fail in a single run.

Split deferrable tasks into scheduled tasks that run at a set time, and
deferred tasks that are deferred indefinitely until they are advanced.
Deferred tasks are more efficient as they live in an unordered hashtable
rather than the sorted list that is used for scheduled tasks.

ComplexTask derived tasks now have a built-in dependency system
instead of needing MTask and its complex queue-within-a-queue
scheme. These tasks are automatically deferred until all dependencies
are satisfied.