Destroy thread args when thread function exits
The args are no longer useful once the thread procedure is finished.
The thread object itself may live longer if the thread's creator holds
onto a reference to check return code, etc. This change makes sure the
args are destroyed as soon as the thread function exits, instead of
waiting for the thread object to be destroyed.