Simplify ComplexTask lastq management
The existing implementation really needed to be atomics to be fully
safe. However upon examaning the requirements, it turns out that it
once a task is added to a queue, it can be guaranteed that it is never
added to another queue until it completes and is reset.

Given this, the logic can be significantly simplified. This not only
removes the need for atomics but also completely avoids a race
condition that previously required a workaround to solve.