Remove taskqueue worker failsafe timer
This used to be needed as there were some small races related to event
signaling. Those have all been fixed and extensive testing has shown
that the failsafe is no longer needed.

Eliminating the failsafe reduces the overhead of having an idle queue
with a lot of worker threads; as they won't wake themselves up
periodically.