Remove conditions on ticking taskqueue manager
In-worker managers only tick if if there's a request for them. In
practice this is almost always, because completed tasks request the
manager to clean them up, and in-worker managers run after task
completions.

Removing this allows us to remove some complexity from the queue
implementation, and also makes the manager more responsive to detecting
when it needs to grow the thread pool in some circumstances.