Add lua
This is the beginning of cxlua, a module intended to provide a lua
scripting environment with additions to make it convenient to use in
a cx-based project.

At this stage it includes the lua interpreter itself (with some
base enhancements, outlined below) and luacmd, a simple command-line
tool for running a lua script, perhaps as part of a build process.

Differences between this interpreter and stock lua are:
* The standard lua libraries have been broken into "safe" and "unsafe"
  categories.
* luaL_openlibs_unsafe is equivalent to luaL_openlibs in a stock lua
  build and gives full access to all standard lua libraries.
* luaL_openlibs on the other hand, creates a safe environment suitable
  for running untrusted scripts in a sandbox. The safe environment:
  * Omits certain core functions - dofile and loadfile
  * Loads a safe version of the os library. This library does not
    include execute, exit, remove, rename, setlocale, or tmpname.
  * Does not include the package, io, or debug libraries, as they are
    inherently unsafe.
* Dynamic library loading is not available on any platform for security
  reasons.