Add cancel cascade
Add flags to ComplexTask that can be set. The first enables a cancel
request to be automatically cascaded to any tasks that the cancelled
task depends on.

This isn't suitable for all circumstances but can be useful, especially
for tasks that create one or more sub-tasks and then depend on them
as a means of waiting for them to complete.