Prevent unnecessary recursive worker wakeups
When a worker thread in a thread pool configuration finishes with a
task, it moves it to the doneq and signals the manager. However, with
in-worker managers, this ends up waking up another worker that has
nothing to do. This is entirely unnecessary as the worker that ran the
task is about to tick the manager anyway.

Add an extra parameter to the manager notify function that allows
this case to set the flag that the worker needs to run, but not
actually wake up the thread. This parameter is still always set to
true for the dedicated manager thread case.