Remove extra worker wakeups
When the thread pool worker finished a cycle, if it successfully
processed any tasks it would wake up another worker at the end,
under the theory that this would help kickstart the queue and deal
with sudden bursts of activity.

That turned out to be completely wrong and waking up another worker
was counterproductive, causing contention storms on the queue. It
also was unnecessary to address the problem, as when large numbers
of tasks are added, new threads already get woken up for each one.