Fix TRMutex sometimes getting stuck
Address a race condition with TRMutex tries to wake up the next task
in the waitlist, but that task has already been woken up and is
currently trying to acquire the mutex, and eventually fails.

This can be caused when many new tasks are being added to the queue
that require the same TRMutex and are consistently beating the task
at the head of the waitlist to the mutex. Eventually those tasks run
out and the missed wakeup causes a stall.

Fix it by checking to see if the task was successfully added to the
advance list (it would fail if the task is not currently in a deferred
state), and advance the next task on the waitlist instead.

This can cause some slight ordering inconsistenty when under load,
but the task run order for TRMutex is already documented to not
be guaranteed.