Handling advancing a task that is being processed
This closes the same race as the previous commit but from the other
side. If a task is advanced while it's concurrently being processed
by the manager and would ultimately be deferred because the requirements
are not yet met, increase a counter to detect this and immediately put
the task back in the advance queue to be re-checked.

Doing it in both places ensure that this doesn't affect TRFifo and
other resource requirements, even though it's been mitigated by
TRMutex.