Correctness for re-advance path in processExtra
The intended course of action in this scenario is to run the entire
manager cycle again. Putting the task back in advanceq within
the loop would cause this to not happen is the task would just be
popped again during the same loop.

Use a temporary array to hold the tasks that need to be re-advanced
and push them into advanceq once the loop completes.