Fix race condition in vfsDestroy
This fixes two issus that can happen when vfsDestroy is called
conurrently with other operations on that VFS.

The first is unexpected destruction of VFSMount objects that are in
use. This is fixed by changing _vfsFindMount to acquire a reference
while vfsdlock is held. This changes the contract of _vfsFindMount:
callers are now required to release the acquired reference when they
are finished with the mount.

The second is failure to completely clear the cache, which can result
in stale references to mounts that no longer exist. Normal vfsUnmount
correctly invalidates the cache, but vfsDestroy does not.