Add timeout support for task requires
This allows a task that has requirements to define an expiration time
after which to give up on the requirement and consider it failed.

The expiration timer leverages the existing task scheduling mechanism
to have the queue recheck the requirements at a specified time.

The task API exposes this capability in the form of variants of the
requirement convenience functions that accept a (relative) timeout.